Count number of in/out msgs in shard connection

+++ router/server/server.go
@@ -33,5 +33,5 @@ type Server interface {
 	Cancel() error
 
 	Reset() error
-	Sync() int
+	Sync() int64
 }
